change publishInstall to run async
Summary: it is not clear from the function definition that it cannot be called on the UI thread.  We have renamed the API in the new SDK to publishInstall[Async|AndWait], but need to make this work until that releases.

Test Plan: Call publishInstall on a UI thread.  see that it works.

Reviewers: mmarucheck, clang

CC: neko-dev@

Differential Revision: https://phabricator.fb.com/D622746