[android-sdk] In PickerFragment, don't reload (unless forced) as long as data is loading, even if it has not loaded.
Summary:
PickerFragment.loadData allowed the caller to specify that a reload should not be forced if data was already present, but
this did not take into account the case where a request had been issued but no response received. In the case of longer
queries, this could result in, e.g., frequent device rotations, constantly restarting the query (and forgetting about the
old one) and never receiving results. Fixed to take into account pending-but-not-finished requests, and also remember
activity circle state between activity restarts, so the activity circle properly reflects the current state.

Test Plan:
- Introduced artificially long delay in async requests (e.g., 10 seconds), repeatedly rotated device
- Observed that activity circle was correct and request completed in the expected time; without change, request
was constantly restarted and did not finish until device rotation ceased.

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D623982

Task ID: 1771196