Change Settings.publishInstall() to be nonblocking (and safe to call in the UI thread), and add publishInstallAndWait() for the blocking case.
Summary: It's not clear from the existing function that it blocks on network requests (and therefore does not work on the UI thread).  This takes care of that.

Test Plan: trigger publishInstall from a UI thread, and verify that it works OK.  attempt to use publishInstallAndWait from the UI thread, and verify that it fails.

Reviewers: mmarucheck, clang

Reviewed By: mmarucheck

CC: neko-dev@, vijaye

Differential Revision: https://phabricator.fb.com/D622381