[android-sdk] Changes to Scrumptious to make it more testable.
Summary:
Writing acceptance tests for Scrumptious is easier if the default text for "What are you eating?" and "Where are you?"
are not the same string. It is also a little more user-friendly. Changed from "Select one" in both cases to "Select a meal"
and "Select a place".

Test Plan:
- Inspection

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D626711