[android-sdk] Added session status callback to LoginButton and LoginFragment.
Summary:
Apps using LoginButton and LoginFragment could benefit from a simple way to add a callback to receive status changes
for the session associated with the button/fragment. Added a get/setSessionStatusCallback to each of them.

Modified BooleanOGSample to demonstrate this alternative approach for receiving session status changes for activities
that are not FacebookActivities.

Fixed LoginButtonTest.testLoginButton to perform necessary actions on the UI thread rather than the test thread.

Test Plan:
- Ran LoginButtonTest
- Ran modified BooleanOGSample to verify that LoginFragment callback also works

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

CC: ekoneil, caabernathy

Differential Revision: https://phabricator.fb.com/D627997

Task ID: 1857226