Update Hackbook profile picture fetch
Summary: October's breaking changes involves the /picture connection, which now returns a dictionary rather than just the URL to the profile picture when the picture field is part of a larger JSON response.  This diff addresses that change so that profile pictures appear again on Hackbook.

Test Plan: Run Hackbook without this code, observe no profile picture is displayed.  Run Hackbook with this code, observe the profile picture.

Reviewers: mmarucheck, clang

Reviewed By: clang

Differential Revision: https://phabricator.fb.com/D629741