[android-sdk] Moved GraphObjectWrapper to GraphObject.Factory.
Summary:
GraphObjectWrapper had a name that did not clearly indicate its purpose. It is really a factory for GraphObjects,
and as such it makes sense to nest it within GraphObject and call it Factory. This commit does that, and also shortens
some of its method names, since it is clear from context now what they are creating.

Fixed a warning in WebDialog.java.

Test Plan:
- Built all projects
- Ran unit tests

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mingfli

CC: ekoneil, caabernathy, platform-diffs@lists, security-diffs@lists

Differential Revision: https://phabricator.fb.com/D632163