[android-sdk] Fix handling of closed sessions in Request.
Summary:
If a Session that is in an unopened state is used in a Request, we were attempting to extract the (null) access token
from it and use it as a parameter. This is an error case, as unopened Sessions should not be used in a request. Moreover,
it was triggering an unclear error message that referenced photo uploads.

Throw a new exception if a Request is executed with a non-null but unopened Session.

Fixed the error message to be more appropriate to the error case.

When a Session is closed, reset the tokenInfo to the same empty token as a newly-created Session has, rather than null,
for consistency.

Test Plan:
- Added new unit test
- Ran all unit tests

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D636352

Task ID: 1901835, 1903573