Using LoginActivity to proxy both SSO and webview login mechanisms.
Summary: This moves most of the logic of handling login behaviors to LoginActivity, and allows us to provide a better exception (in the app's process) to fail fast if the calling activity is singleInstance.

Test Plan: Ran unit tests, tested with Scrumptious and SessionLoginSample (on different login behaviors), as well as Hackbook for back compat.

Reviewers: mmarucheck, clang

Reviewed By: clang

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D638002

Task ID: 1901717