Update facebook/src/com/facebook/android/Facebook.java
initialise mAuthActivityCode with DEFAULT_AUTH_ACTIVITY_CODE to prevent NPE when launching an non-facebook activity with startActivityForResult and a requestCode to 0. In that case, if facebook code is present in the onActivityResult, it will be launched and mAuthActivityCode (=0 because not initialised) will be compared to requestCode (=0) then null objects will be called.