[android-sdk] Makes AccessToken public, adds Session.open that takes an already-created AccessToken.
Summary:
Opening a Session using an access token provided by the Facebook application via a native-linking Intent is a common scenario, but was not straightforward to implement. This commit adds methods to AccessToken to simplify handling such Intents.

AccessToken is now a public class, and there is a Session.open overload that takes an AccessToken and opens a session using it.

Added sample code to BooleanOGSample to show a more sophisticated handling of native linking.

Added a NonCachingTokenCachingStrategy for use in scenarios where no tokens should be cached (for instance, temporarily opening a session to handle a deep-linking request).

Test Plan:
- Added unit tests, ran them
- Verified that deep linking works in BooleanOGSample

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks, chrisp

Reviewed By: mmarucheck

CC: jacl

Differential Revision: https://phabricator.fb.com/D640046

Task ID: 1914904