[android-sdk] Adds RequestBatch-level completion callback; fixes Session reauth behavior.
Summary:
Some scenarios involving batched Graph API calls are more cleanly implemented with a completion callback that applies
to the batch as a whole, rather than individual requests within the batch. This commit adds such a callback to
RequestBatch.

It also fixes a couple issues with Session and reauthorization. It adds a check to ensure that the access token resulting
from a reauthorization represents the same Facebook profile as the previous access token did (since users could authenticate
using different credentials than when the session was originally opened). It also fetches a current list of permissions
for the user following a reauthorization, to help ensure that the client copy of the permission list stays in sync with
the service.

Finally, it fixes LoginActivity to pass back the AccessTokenSource of the access token it provides, so we can set it properly.

Test Plan:
- Added unit tests
- Ran unit tests
- Ran Scrumptious, HelloFacebook, and BooleanOG samples

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

CC: ekoneil, caabernathy

Differential Revision: https://phabricator.fb.com/D647716

Task ID: 1925038, 1925028