Added note to Facebook.java discouraging its use
Summary:
Making the entire Facebook class deprecated causes compiler warnings
in other deprecated classes.  As an alternative, we have deprecated
all public members.

Based on feedback, this augments the docstring to make this more
obvious and adds an explanation of why the class is not deprecated.

Test Plan:
- build

Revert Plan:
-safe

Reviewers: ekoneil, mingfli, clang

Reviewed By: clang

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D648641

Task ID: 1938802