Fix issue with friends not showing up on restart.
Summary: The issue is that between starts on the emulator, the app itself might not have been destroyed, which causes the Active Session to stick around, which means that on next startup, it starts in the OPENED state, and there's no state change, which normally triggers the load. The fix is to check whether the Active Session is already opened, and to just load the data then and there.

Test Plan: tested on emulator

Reviewers: clang, mmarucheck

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D653861

Task ID: 1903570