Fixing erroneous call to uiHelper in SelectionFragment.
Summary: Fixing the call to be the right method to uiHelper, also adding a check in init() to see if the session is already open, and making the me request there as well (since if it's already open then there won't be a transition).

Test Plan: tested on emulator, ran unit tests.

Reviewers: clang, mmarucheck

Reviewed By: mmarucheck

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D656165

Task ID: 1955361