[android-sdk] Don't give focus to search box immediately when PlacePickerFragment shown; naming issues in SessionLoginSample
Summary:
Previously, the search box (if displayed) in PlacePickerFragment would get focus as soon as the picker was shown, bringing
up the keyboard on many devices and making it difficult to scroll the list of results. Changing to keep it from getting
focus until the user touches it so the entire area is used for displaying results.

SessionLoginSample had a poorly-named MainActivity and referred to the since-renamed LoginFragment in its UI.

Test Plan:
- Ran PlacePickerSample, did not see the keyboard until I pressed in the search box
- Ran SessionLoginSample, verified rewording in UI

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D656709

Task ID: 1956968, 1957867