[android-sdk] Rather than throw if an auth request comes in when one is pending, report the error via the callback and leave state unchanged.
Summary:
Previously, we would throw if an open()-ish call came into a Session while another request was pending. App UI may allow
this to happen (for instance, a user clicking rapidly on a login button that doesn't disable itself while a request is
pending), a cleaner solution would be to handle the new request as if it was a failed reauth and call the state-change
handler, leaving the state unchanged but providing an Exception so developers can handle/log.

Test Plan:
- Ran SessionLoginSample, clicked rapidly on Login button in UserSettingsFragment UI, did not see crash (but did see
Exception being reported via postStateChange)
- Added unit test, ran unit tests

Revert Plan:

Reviewers: mmarucheck, mingfli, karthiks

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D656843

Task ID: 1957947