Fix memory leak in LoginManager
LoginManager leaked the calling activity by referencing it with context
and loginLogger. This was only fixed in case of a successful login, but
not when the login was canceled. Actually there is no need to hold a
reference to the calling activity as a reference to the application
context is sufficient.