Fix memory leak in LoginManager and AppEventsLogger
LoginManager's context and loginLogger both hold a reference to the
activity starting the login. This activity is leaked if the login process
is canceled. It is also leaked if finishLogin() is never called, e.g. when
the device's configuration/orientation is changed. finishLogin() only nulls
out context and loginLogger in case of a successful login.

This commit updates LoginManager and AppEventsLogger, so that it is not
required anymore to hold a reference to the activity starting the login.