Remove unnecessary volatile keyword
The volatile keyword is only needed when a field is lazily initialized with double-check idiom. Some of the classes in this sdk uses such idiom - for example DeviceLoginManager class. In double-check idiom, it is critical for a field to be declared volatile, because the first check is performed without locking. But when synchronized accessor is used, volatile keyword is not needed anymore. All classes that I removed volatile keyword already use synchronized keyword. Unnecessary volatile keyword may increase runtime overhead, and decrease code readability.

Also I think RequestState and SessionInfo classes should be thread-safe by becoming immutable or synchronized. For example, DeviceAuthDialog class has a volatile field "currentRequestState", but declaring a reference field as volatile does not ensure visibility of non-volatile fields that are accessed via that reference. None of RequestState class' fields are declared as volatile. As Brian Goetz mentioned "in the absence of synchronization, it is possible to see an up-to-date value for an object reference that was written by another thread and still see stale values for that object's state." (https://www.ibm.com/developerworks/java/library/j-jtp06197/) SessionInfo is the same, which is used in ActivityLifecycleTracker class.