Use the expires_in field to pass into access token class
Summary: For some reason, we aren't doing this so client thinks it is an infinite token. We should return the correct time. I really don't like expires_in, and feel expiration_time is better so we don't need to do the logic shift twice, and just return the raw field

Reviewed By: jingping2015

Differential Revision: D10045942

fbshipit-source-id: a730b8b139c03a28b6a4cc11f32b05567793b1a2