Make network dependant calls non-synchronized
Summary:
Network calls can take very long if network is not available
and network timeout is long. Once synchronized method call holds the
lock for too long, then the other synchronized call will be blocked
until that lock is released. As far as we are protecting the data
from multiple threads, synchronization as method level is not needed.

Reviewed By: khp

Differential Revision: D10437143

fbshipit-source-id: c254e8159ff1b05897485d3be32ac2c8673a128b