Silently catch exceptions from invalid enum values (#544)
Summary:
Your current SDK crashes all the time for real users:
```
java.lang.IllegalArgumentException:
   at java.lang.Enum.valueOf (Enum.java:257)
   at com.facebook.appevents.codeless.internal.EventBinding$MappingMethod.valueOf (EventBinding.java:154)
 at com.facebook.appevents.codeless.internal.EventBinding.getInstanceFromJson (EventBinding.java:82)
  at com.facebook.appevents.codeless.internal.EventBinding.parseArray (EventBinding.java:69)
  at com.facebook.appevents.codeless.CodelessMatcher$ViewMatcher.run (CodelessMatcher.java:224)
  at android.os.Handler.handleCallback (Handler.java:789)
  at android.os.Handler.dispatchMessage (Handler.java:98)
  at android.os.Looper.loop (Looper.java:164)
  at android.app.ActivityThread.main (ActivityThread.java:6944)
  at java.lang.reflect.Method.invoke (Native Method)
  at com.android.internal.os.Zygote$MethodAndArgsCaller.run (Zygote.java:327)
  at com.android.internal.os.ZygoteInit.main (ZygoteInit.java:1374)
```

There already is a bug report for this but it has not been worked on yet because it is stuck in "need more info" state:
https://developers.facebook.com/support/bugs/278547559523105/

Looks like this comes from invalid/unexpected values delivered by your codeless configuration JSON data.

What this change does
--
This change fixes the uncaught exceptions by catching them to avoid the resulting crashes. Imho this should be done anyway for incoming data processing and should have been done from the start of this feature.

What this change doesn't do
--
The unexpected values in the configuration data have to be handled somehow of course but this is out of the scope of this fix.

Misc
--
To whom it may concern: your pull request template seems to be out of date. It tells me to issue my pull request against your `:dev` branch but there is no such branch.
Pull Request resolved: https://github.com/facebook/facebook-android-sdk/pull/544

Reviewed By: wi3lau

Differential Revision: D13948378

Pulled By: jingping2015

fbshipit-source-id: 2bbe6fd40dad62360f0eec3f27e00d46b284c580