Fix issue that stop sending install events after a failed request
Summary:
`lastPing` will be set when there is no exception from graph request, however, graph request will catch exceptions, so the `lastPing` will be set no matter the request succeeded or not, which is not ideal since there is chance that network is unavailable or graph API breaks.

So we should check whether install event has been successfully sent before set the `lastPing`.

Reviewed By: tianqibt

Differential Revision: D16609854

fbshipit-source-id: bf7b33c2cd8790c9207a0944334d4ce836891db5