Install referrer uses new API
Summary:
InstallBroadcast will be deprecated in March, 2020 according to https://android-developers.googleblog.com/2019/11/still-using-installbroadcast-switch-to.html hence we need to update to use the new API.

This information needs only to be retrieved once.
Reason: Google does not change it for 90 days. During app uninstall (assuming no backup) the sharedpreferences are deleted as well, and the data can not possibly change when the app is not re-installed.

A small nit-problem is since we are making an async call to retrieve the data, the first time we get the URL it might not be there in time so we will receive null. As explained in the comments it should not be a problem, as all subsequent calls will have it.

Reviewed By: joesus

Differential Revision: D19237948

fbshipit-source-id: 4f4cc040d882c585918d870acd60da865278760b