Startup metrics API
Summary:
The idea is that API will be used internally to measure different possible startup metrics starting with cold start.

Note that TAG does not necessarily have to be 1:1 mapping to an SDK folder. The idea is that the TAG can describe your use-case. Since this will be internal API we have freedom to change/refactor naming of the tags if the needs arise.

The usage would be (after TAG enum is moved to its own file).

```
MetricsUtil.getInstance(applicationContext).startMeasureFor(FACEBOOK_CORE_STARTUP);
//your code
MetricsUtil.getInstance(applicationContext).stopMeasureFor(FACEBOOK_CORE_STARTUP);
```
AppEventsLoggerImpl can later poll this data with
```
MetricsUtil.getInstance(applicationContext).getLastTimeDifferenceFor(FACEBOOK_CORE_STARTUP);
```

Even if we decide to directly send it to the server, we would need a service that checks that SDK is fully initialized and then sends it separately. We would still need to store the data somewhere on device (sharedpreferences). I prefer AppEventsLoggerImpl approach because we need to map the startup time to some appid+device anyway which is handled by AppEvents already.

If adding new metrics to measure, it would be about adding new HashMap<Tag, YOUR_METRIC> and a new base string for that metric. The start/stop measure would collect all metrics, and then it will be up to the user  to query for the data they want.

The goal is to make an easy to use API.
Are there some use-cases that will not be covered by this?
For example, the class is not really thread-safe but it also does not make sense to call startMeasure/endMeasure for same tag from different threads.

All kinds of feedback is appreciated.

Reviewed By: jingping2015

Differential Revision: D19318122

fbshipit-source-id: 840a00c4da4b8ed4ddc0df6a48d7f39e66e8ad35