Deletes LoginSample sample app
Summary:
The LoginSample shows how to use the login- and accountkit-sdk.
Reasons for removal:
FBLoginSample already shows how to use the login functionality, and accountkit will be deprecated on 9 March, 2020 (https://developers.facebook.com/docs/accountkit/android)
Hence it is better to remove sample apps so developers do not use it as example code.

Differential Revision: D19499950

fbshipit-source-id: 7f9e669df2b39b53371d690d4c1b6efcfe470675