Deletes LoginSample sample app (#656)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-android-sdk/pull/656

The LoginSample shows how to use the login- and accountkit-sdk.
Reasons for removal:
FBLoginSample already shows how to use the login functionality, and accountkit will be deprecated on 9 March, 2020 (https://developers.facebook.com/docs/accountkit/android)
Hence it is better to remove sample apps so developers do not use it as example code.

Reviewed By: joesus

Differential Revision: D19499950

fbshipit-source-id: 2bf19fb1c2201b63dba2e4b20aa18c4840d050f5