Asks developer to use their app id for sample apps
Summary:
Title.
There are a few reasons:
1) We do not want developers to use app ids of sample apps for their real apps.
2) We want to understand how useful the sample apps are.

This requires the devs to follow the getting-started guide.

Bypasses alert if you change the string to anything else, but giving a non-valid appid will not make it work correctly.

Only doing it for FBLoginSample app for now. After we decide what other sample apps to keep we can do same thing there.

Differential Revision: D19539088

fbshipit-source-id: 8f30e694a6012e2e9ad18948a90e469eacbc1770