Add MonitorLoggingStore
Summary:
MonitorLoggingStore will read/write the logs from/to the disk.
We will delete the logs in the disk once they are read from the disk in order to avoid sending them twice.

The logs in the LoggingCache will be written into the disk only when the app is stopped/crash. We will flush the disk once we enable the monitor.
We will discard the logs if they failed to be sent, for the reason 1) our log is not that important 2) reduce the usage of storage.

Reviewed By: Mxiim

Differential Revision: D20477932

fbshipit-source-id: 523bf5cb36ce5bfe3240730be33bb33ab72801bf