Remove MonitorEvent class and modify MonitorLog (#749)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-android-sdk/pull/749

Since we added category attribute, and decided to track unique application id which is the package name for Android, I added these new fields in `MonitorLog` class.

`MonitorLog` has a required attribute `LogEvent` which contains event name and category. If the category is `PERFORMANCE`, the event name should be one of the values in `validPerformanceEventNames` Set.

Also, I removed `MonitorEvent`, because we want to track the usage of method in the future. The method name will be stored under event name field, using pattern className::methodName. We can't make the `MonitorEvent` as an Enum any more.

Reviewed By: Mxiim

Differential Revision: D20775666

fbshipit-source-id: 36de5fcf14f6cf68943a9d5645b486ab4bd0f1a9