We androidX now (#766)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-android-sdk/pull/766

Migrating to android X for next major release on 5th May.
Reason being is a few partners are asking for it, and it is nice to use shiny new things.

At the same time updating travis to use android x sdk.

I plan to rebase after our release on 20th April, but want to hear your opinions and concerns now.

Reviewed By: jingping2015

Differential Revision: D20074358

fbshipit-source-id: 13bdc460dbc84a188d1d85915605ccd5d9763be8