Enable Monitor during SDK is initializing (#779)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-android-sdk/pull/779

Added MonitorManager to control enable or not the Monitor
Change loadSamplingRatesMapAsync() from public to protected (allow test to access)
Enabled Monitor when SDK is initializing

The dev need to turn on the feature toggle on Dev site to enable Monitor Feature.
To support the EU consent, the dev can set MonitorEnabled to false in AndroidManifest and once they get user's consent call Facebook.setMonitorEnabled(true) to re-enable.

The default value of the flag MonitorEnabled in AndroidManifest is true.

Reviewed By: Mxiim

Differential Revision: D21506564

fbshipit-source-id: 4e53a24492fe2778ad2f9213e6ba483ed28c2658