Create fuzzy test for Android SDK setting fetch
Summary:
As SEV follow up, we want to make sure that wrong server-side data doesn't cause SDK initialization to break.

I'm introducing new test helper class that takes JSON string and tries to replace any key with random type. SDK should never break for all of this inputs.

Due to my lack of knowledge on Android testing or/and old testing frameworks, I had to come up with pretty awful code.
- Robolectric 3 explodes when JSON(or just Object) types are being passed. Migrating to Robolectric 4 is different big thing.
- JSON package is unmocked and not allowed to be called in preparation methods, forcing me to write yet another awful code. I'm truly sorry.

But in the end it works, it generates ~100 tests with different variations of input JSON.

Reviewed By: dreamolight

Differential Revision: D21560174

fbshipit-source-id: 7186047df4379772c6c7f9871310eddec8d738cb