share kit
Summary:
Added to class-level, because only methods that throw are private and used within the same class, they are also properly caught meaning that the crashhandler should not affect them anyway.
Left out `DeviceShareDialog` since it appears it needs to throw
Left out `VideoUploader` since it looks like the methods are either throwing exception manually, or require the caller to take care of it, or have callbacks for errors. We can add to it if we get reports that it crashes here.

Reviewed By: jingping2015

Differential Revision: D22142359

fbshipit-source-id: 2f6c2d0c0bee93c8b11aaa264b3790d9862b3004