prepare imagerequest for apiv8
Summary:
title, similar to D22699837.
calls to profile need to have token, if token is null it will at least return silhoutte picture. This case shouldnt happen since you will probably be logged in already.

This is a special case when we do not through GraphRequest.
Inside GraphRequest even if accesstoken is null, we will check of clienttoken. Later on, we can add similar check to graphRequest.

Reviewed By: Nathaaaalie

Differential Revision: D22802750

fbshipit-source-id: f9afbfd0e44142c58f0865ca94bd3e91438b7d32