adding necessary classes for qvt
Summary: needed for rest of the stack, because the "original" annotations are outside opensource and our repo. We _could_ include them in buck, but not in gradle. Hence we create them here. For the purposes how they are used, the naming of them is the important thing, and not their import.

Reviewed By: Nathaaaalie

Differential Revision: D23551493

fbshipit-source-id: 5aff2e733bfe3025aab2dfa6325085116366021b