Fix crash in device login success
Summary: This one is super hard to debug and repro. I can ocasionally repro it, but it usually works. When I see a crash log, I see that the following class variable "deviceAuthMethodHandler" (in comments) is null. So simply null checking should works. I believe this happens due to race condition between the short polling and the dialog dismissal.

Reviewed By: Mxiim

Differential Revision: D23892194

fbshipit-source-id: 4954b16d3058bb49e5d1afa8fa75b5f2e5c3110d