Fix orientation change bug
Summary:
We actually want to use onDestroyView to set isBeingDestroyed = true because onDestroyView is being called before onDismiss when orientation change happens. In this case, we do not actaully cancel the dialog.

When user dismissed the dialog, onDismiss is called first

Now, when orientation change happens, we keep the login request so that it can persist throughout the orientation change

Reviewed By: Mxiim

Differential Revision: D23913834

fbshipit-source-id: 54824b4e1d6a7ea0a1bc9a6f0b322e8bc49e8acf