Do not check if intent exists before firing proxy auth intent
Summary:
Building with Android 11 affects package visibility. Reference: https://developer.android.com/about/versions/11/privacy/package-visibility.

To mitigate, we will try calling the intent and hope it works as package visibility doesn't limit intents firing to other apps. We now need to change tryAuthorize to return an int as it will return the amount of intents that it fires.

Reviewed By: Mxiim

Differential Revision: D24000953

fbshipit-source-id: db13043e17d5e5f17d15d841bcba9a63568a4ab1