Launch Custom tab intent without querying if it exists
Summary:
Building with Android 11 affects package visibility. Reference: https://developer.android.com/about/versions/11/privacy/package-visibility.
For Custom tab details for Android 11, see this guide: https://developers.google.com/web/android/custom-tabs/implementation-guide#how_can_i_check_whether_the_android_device_has_a_browser_that_supports_custom_tab

Notice that the app needs to add the intent to the queries element.

This diff changes the way FB Login on Android handles Custom Tabs. Instead of initially querying for the package and only trying to open Custom tabs if it does succeed, now this will attempt to open and catch ActivityNotFoundException if it does not succeed. It then falls back to the next handler in that case.

The functionality changes a bit when building with 30. It defaults to choosing the user's preferred browser. Originally, it would pick the first intent it finds, which may not be the user's preferred browser.

Reviewed By: regresscheck

Differential Revision: D24096036

fbshipit-source-id: 9edeb4b876352e0a9fd5649379ee25df124a3a7f