Only keep track of numActivitiesReturned for KatanaProxyLoginMethodHandler
Summary: In KatanaProxyLoginMethodHandler, we can potentially fire multiple intents, depending on which FB apps are installed on the device. As a result, we would need to keep track of how many might return. However, for other login methods such as CustomTabLoginMethodHandler, we will only fire 1 potentially, so there is no point in checking logic for numActivitiesReturned.

Reviewed By: jvlyn

Differential Revision: D24257099

fbshipit-source-id: 97f71442a272db9f3341421af4a5508829e63ee8