Fix CodelessManagerTest failure on Travis CI (#849)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-android-sdk/pull/849

The failure can not repro locally, I just narrowed the cause area down to the mock of the static function checkCodelessSession. For some reason it doesn't work on Travis CI. The reason of mocking this is to verify it has been called under some test cases, since Robolectric4 stop supporting verify a particular static function has been called.

Refactor CodelessManager so that makes it testable.

Reviewed By: Mxiim

Differential Revision: D24482448

fbshipit-source-id: 8c4c6341586fd02dd56c1cd0042d04e46e950d9f