logging change setting implicitly using both current and previous values of autlogEventsEnabled
Summary:
# Context
When changing the setting for autoLogAppEventsEnabled from true to false there is not log because the logging is made with the actual value of autoLogAppEventsEnabled as disabled.
we need to log this case anyway; in order to know that will be not further events logging.
The CPIR authorising  this modification is
https://fburl.com/launch_manager/ml2ft18y

Reviewed By: dreamolight

Differential Revision: D24711002

fbshipit-source-id: c8e7d977a25e4151d9baa091db4659665e48ea01