Add MTensorTest.kt to test MTensor.java
Summary:
Added MTensorTest.kt for testing MTensor.java.
There was an existing test for testing the reshape method in MTensor in OperatorTest.java. Moved it out of OperatorTest.java and into MTensorTest.kt. Also added tests for getShape() and getShapeSize().

Also updated the `for` loop in MTensor.java since there was an Android Studio suggestion for it.

Reviewed By: Mxiim

Differential Revision: D25012522

fbshipit-source-id: eea4da10bf00c3441f72d7eae0ab2d61f40e7f05