remove travis completely
Summary:
See task
We are already running github actions, ran in parallel to make sure no issues. Now time to kill travis.
As far as release goes, the person releasing would need to manually check it is passing (as before), alternatively we could look into githubs API what they have, but this is out of scope of this diff. Additionally, the right solution is to release from github actions directly, but we cannot do that due InjKit.

Reviewed By: jingping2015

Differential Revision: D25589466

fbshipit-source-id: 567ef52f7af7dfe7e974b1071962201d1564cde2