Removes two deadlocks
Summary:
almost tl;dr
FacebookSDK.java has a lot synchronized methods,
when it calls statically class A or class B, they call methods from FacebookSDK themselves, but the lock on those methods is acquired already because it is synchronized on class lock.
This is a ticking bomb. Right now it worked because we call things in a very specific order and call it in Executor first time.
Hence the solution is to
1) remove class lock and have variable lock on initialized. This should be safe because we just change what we lock on.

2) remove dependency on another synchronized method since it is not needed. This should be safe as the SDK should already be initialized to get applicationID. This is either done in ContentProvider automatically, or will be done in your Application/Main-activity as recommended in the docs. If there are concerns, we can always add same method unsynchronized to only be used inside AppEventsLogger since the possible deadlock can arise from there.

Reviewed By: jingping2015

Differential Revision: D26157231

fbshipit-source-id: 1c73d4fcfcd375678a590b057a25f0354cae2b6b