Re-enable several more unit tests
Summary:
These tests have been disabled for a while. They all seem to be passing though? (Not sure what the original rationale was behind disabling them - seemed to be related to some appevents changes.)

There was only one test I wasn't able to re-enable and that's `ExceptionAnalyzerTest` because it relies on `!BuildConfig.DEBUG`. We could technically make this work by setting up a different build config for unit tests but that seems like a lot of work to make one unit test out of 700+ work so I'm leaving the `Ignore` for now.

Reviewed By: Mxiim

Differential Revision: D26442339

fbshipit-source-id: 85b445f3d25e3b1a76617ee8c85b638c730aa9c8