Add UrlRedirectCacheTest and fix UrlRedirectCache on loop redirects
Summary:
This is the test for `UrlRedirectCache`. 3 static methods are tested in this unit test:
1. `cacheUriRedirect(Uri, Uri)`
2. `getRedirectedUri(Uri)`
3. `clearCache()`

This diff also fixes UrlRedirectCache if there is a loop in the redirect chain. If there is an A->A loop, the cache will return A. If there is more than or equal to 2 addresses in the loop, null will be returned and an error will be reported in log.

Reviewed By: joesus

Differential Revision: D26624011

fbshipit-source-id: 15728a4c75f1f1e02db4e7c8c406c9e37fe2f99d