CallbackManagerImpl
Summary:
Move `CallbackManagerImpl` into Kotlin.

The nullable `Callback` arguments are removed because if the argument is null, an exception will be thrown in the original code. With Kotlin non-nullable argument, it's simpler to keep the safety. Also, test code for checking whether it will throw exceptions when passing a null value is removed.

Reviewed By: jingping2015

Differential Revision: D26583629

fbshipit-source-id: faa92bd8279edf42791fa8f4b4b0afbea99ccb02