Do not close CustomTab when home button is pressed
Summary:
When user has 2FA enabled, there's no way for them to finish Facebook Login as they usually need to switch to another app. This closes current flow, and they are stuck.

We added this line as a mitigation for an intent filter bug from 5 years ago. This doesn't seem to reproduce nowadays. Should be safe to remove.

Reviewed By: jvlyn

Differential Revision: D27266684

fbshipit-source-id: 63d0e211ca2dc514baa3c3fedc6d07d7c9010874