Migrate AccessToken into Kotlin
Summary:
As title. `AccessToken` is an important and foundational class in the SDK, hence I try my best to keep the API unchanged. However, we still have very small differences in the behaviors. I believe these differences won't lead to a change in client codes (in fact, we only make client code easier). Here are the differences:
1. `token`, `applicationId`, `userId` and `permissions` fields are promised to be nonnull. The reason is that according to its constructor, these properties must be nonnull if it's properly constructed.
2. The validations on `token`, `applicationId` and `userId` are also enforced when the users call `CREATOR.createFromParcel()`. Now if the source parcel object is invalid, there will be an exception rather than returning an invalid `AccessToken` object. It's reasonable because `createFromParcel()` assumes the parameter is correct and valid.

There is still a problem needs to be done: an interface for callback has terrible function naming. We can only fix it in a major release so I keep it unchanged.

Also, we have some static getters and I keep them rather than using static fields. If we use static fields, the interface to Java is unchanged, but the interface to Kotlin will be changed. It's interesting that Kotlin will only treat getter/setters on objects as fields when import Java code.

Reviewed By: jingping2015, Oliverccccct

Differential Revision: D27975759

fbshipit-source-id: 08a5dd376a0e42790af9f4ce689d3fbff27c3325