Migrate GraphRequest to Kotlin
Summary:
As title. `GraphRequest` is one of the most important class in the SDK and this diff unavoidably changed the external interfaces. I try my best to minimize the changes and for Java code nothing need to be changed. But there still some differences in Kotlin side.

1. `Callback.onComplete()` will have a nonnull parameter.For Kotlin side, if the client code implements it with nullable parameters, it will be treated as different function. It can be fixed with very little effort.
2. `GraphJSONArrayCallback` and `GraphJSONObjectCallback` will have a nullable parameter following the documents. For Kotlin side if the client code implement it with nonnull parameter, it will be treated as different function. It will force the client to do null check on the parameters.

Also updated the tests that require mocking static methods of GraphRequest.

`GraphRequest` in ushba is also migrated to Kotlin, otherwise many tests that needs to mock behaviors of GraphRequest will fail.

When `GraphRequest.java` is in Java, public version and ushba version have following differences:
1. [ushba add] a new private field named `graphDomainOverride`
2. [ushba add] getter/setter for a public field name `graphDomainForSingleRequest`. This field is computed based on `graphDomainOverride`
3. [ushba add] new private methods `getAccessTokenToUseForRequest`, `getClientTokenForRequest`, `getUrlWithGraphPath`, `isValidGraphRequestForDomain`, `shouldForceClientTokenForRequest`, `isApplicationRequest`
4. [ushba modify] implementation of private method `addCommonParameters`

So all added fields/methods are translated to Kotlin in this diff. I also follow the ushba's `addCommonParameters` to write the Kotlin version of it.

Reviewed By: jingping2015

Differential Revision: D28341175

fbshipit-source-id: 503c8171763423d7d486fd29b5fbbaca6dc455a6