Migrate FacebookBroadcastReceiver to Kotlin
Summary:
As title.

Following the semantics, all callback methods in `FacebookBroadcastReceiver` will have nonnull parameters in Kotlin.

Note: I found that for all protected methods declared in Kotlin, they're not allowed to be accessed in test. If we do need to access them, we can inherit the class the make all protected methods public without changing the original visibility settings.

Reviewed By: jingping2015

Differential Revision: D28393069

fbshipit-source-id: 06245f973f50f03f7a4b0a748fdb4f334facf683