Remove dependencies on deprecated kotlin-android-extensions
Summary:
We don't actually use any utilities from `kotlin-android-extensions`. It's easy to remove it from the gradle configuration.

`kotlin-android-extensions` will be removed from Kotlin from this September (https://android-developers.googleblog.com/2020/11/the-future-of-kotlin-android-extensions.html) when Kotlin 1.5 releases. Adding it in the dependency will block developers to upgrade to Kotlin 1.5 in the near future.

Reviewed By: jingping2015

Differential Revision: D28422597

fbshipit-source-id: 502df9ccd86bf38ff59578ca28e716b830ffb386